Moved cnapp services
