+++ src/main/java/org/k8loud/executor/actions/cnapp/db/mysql/MySqlUpdateAction.java
@@ -3,8 +3,8 @@ package org.k8loud.executor.actions.cnapp.db.mysql;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import org.k8loud.executor.actions.cnapp.db.DBAction;
-import org.k8loud.executor.db.DBService;
-import org.k8loud.executor.db.SuperConnection;
+import org.k8loud.executor.cnapp.db.DBService;
+import org.k8loud.executor.cnapp.db.SuperConnection;
 import org.k8loud.executor.exception.ActionException;
 import org.k8loud.executor.exception.CustomException;
 import org.k8loud.executor.exception.DBException;
