+++ src/main/java/org/k8loud/executor/cnapp/db/Operations.java
@@ -1,4 +1,4 @@
-package org.k8loud.executor.db;
+package org.k8loud.executor.cnapp.db;
 
 public enum Operations {
     UPDATE,
