+++ src/main/java/org/k8loud/executor/cnapp/http/HTTPService.java
@@ -1,4 +1,4 @@
-package org.k8loud.executor.http;
+package org.k8loud.executor.cnapp.http;
 
 import org.apache.http.HttpResponse;
