+++ src/main/java/org/k8loud/executor/cnapp/mail/MailProperties.java
@@ -1,4 +1,4 @@
-package org.k8loud.executor.mail;
+package org.k8loud.executor.cnapp.mail;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
