+++ src/main/java/org/k8loud/executor/cnapp/mail/MailService.java
@@ -1,4 +1,4 @@
-package org.k8loud.executor.mail;
+package org.k8loud.executor.cnapp.mail;
 
 import org.k8loud.executor.exception.MailException;
