+++ src/main/java/org/k8loud/executor/cnapp/mail/MailServiceImpl.java
@@ -1,4 +1,4 @@
-package org.k8loud.executor.mail;
+package org.k8loud.executor.cnapp.mail;
 
 import jakarta.activation.FileDataSource;
 import lombok.RequiredArgsConstructor;
