+++ src/main/java/org/k8loud/executor/cnapp/sockshop/SockShopServiceImpl.java
@@ -4,13 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.http.HttpResponse;
 import org.jetbrains.annotations.Nullable;
+import org.k8loud.executor.cnapp.http.HTTPService;
+import org.k8loud.executor.cnapp.http.HTTPSession;
+import org.k8loud.executor.cnapp.mail.MailService;
 import org.k8loud.executor.cnapp.sockshop.params.CreateAddressParams;
 import org.k8loud.executor.cnapp.sockshop.params.RegisterUserParams;
 import org.k8loud.executor.datastorage.DataStorageService;
 import org.k8loud.executor.exception.*;
-import org.k8loud.executor.http.HTTPService;
-import org.k8loud.executor.http.HTTPSession;
-import org.k8loud.executor.mail.MailService;
 import org.k8loud.executor.util.Util;
 import org.k8loud.executor.util.annotation.ThrowExceptionAndLogExecutionTime;
 import org.springframework.stereotype.Service;
