+++ src/main/java/org/k8loud/executor/drools/UsableServices.java
@@ -2,13 +2,13 @@ package org.k8loud.executor.drools;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.k8loud.executor.cnapp.db.MongoService;
+import org.k8loud.executor.cnapp.db.MySQLService;
+import org.k8loud.executor.cnapp.http.HTTPService;
+import org.k8loud.executor.cnapp.mail.MailService;
 import org.k8loud.executor.cnapp.sockshop.SockShopService;
 import org.k8loud.executor.command.CommandExecutionService;
-import org.k8loud.executor.db.MongoService;
-import org.k8loud.executor.db.MySQLService;
-import org.k8loud.executor.http.HTTPService;
 import org.k8loud.executor.kubernetes.KubernetesService;
-import org.k8loud.executor.mail.MailService;
 import org.k8loud.executor.openstack.OpenstackService;
 import org.springframework.stereotype.Component;
