+++ src/main/java/org/k8loud/executor/service/MapperServiceImpl.java
@@ -5,17 +5,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.k8loud.executor.actions.Action;
 import org.k8loud.executor.actions.ActionHelper;
+import org.k8loud.executor.cnapp.db.DBService;
+import org.k8loud.executor.cnapp.db.MongoService;
+import org.k8loud.executor.cnapp.db.MySQLService;
+import org.k8loud.executor.cnapp.http.HTTPService;
+import org.k8loud.executor.cnapp.mail.MailService;
 import org.k8loud.executor.cnapp.sockshop.SockShopService;
 import org.k8loud.executor.command.CommandExecutionService;
 import org.k8loud.executor.datastorage.DataStorageService;
-import org.k8loud.executor.db.DBService;
-import org.k8loud.executor.db.MongoService;
-import org.k8loud.executor.db.MySQLService;
 import org.k8loud.executor.exception.ActionException;
 import org.k8loud.executor.exception.MapperException;
-import org.k8loud.executor.http.HTTPService;
 import org.k8loud.executor.kubernetes.KubernetesService;
-import org.k8loud.executor.mail.MailService;
 import org.k8loud.executor.model.ExecutionRQ;
 import org.k8loud.executor.model.Params;
 import org.k8loud.executor.openstack.OpenstackService;
