+++ src/test/java/org/k8loud/executor/http/HTTPSessionTest.java
@@ -13,9 +13,9 @@ import org.apache.http.message.BasicStatusLine;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.k8loud.executor.cnapp.http.HTTPSession;
 import org.k8loud.executor.cnapp.sockshop.params.RegisterUserParams;
 import org.k8loud.executor.exception.HTTPException;
-import org.k8loud.executor.http.HTTPSession;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Spy;
